Add Alpine 3.10, Debian 10, Debian Docker

+++ scripts/cleanup_linux.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-rm /etc/hosts
+[ -f "/etc/hosts" ] && rm /etc/hosts
 if [ $PACKER_BUILD_NAME != 'docker' ]; then
 	cloud-init clean
 fi;
